Fix signed/unsigned warning.

+++ mux/src/htab.cpp
@@ -1,6 +1,6 @@
 // htab.cpp -- Table hashing routines.
 //
-// $Id: htab.cpp,v 1.12 2004-07-07 17:00:47 sdennis Exp $
+// $Id: htab.cpp,v 1.13 2004-07-07 18:00:08 sdennis Exp $
 //
 // MUX 2.4
 // Copyright (C) 1998 through 2004 Solid Vertical Domains, Ltd. All
@@ -50,7 +50,7 @@ void *hashfindLEN(const void *str, size_t nStr, CHashTable *htab)
     {
         HP_HEAPLENGTH nRecord;
         htab->Copy(iDir, &nRecord, &htab_rec);
-        int nTarget = nRecord - sizeof(int *);
+        size_t nTarget = nRecord - sizeof(int *);
 
         if (  nTarget == nStr
            && memcmp(str, htab_rec.aTarget, nStr) == 0)
